requirements.txt
setuptools==80.9.0
Pillow==9.5.0
numpy==1.26.4
pandas==2.3.3
streamlit==1.27.0
streamlit-drawable-canvas==0.9.3
torch==2.9.1
torchvision==0.24.1
